chore(test): clean up tests in non-Saluki crates (#2139)
