+++ lib/stringtheory/Cargo.toml
@@ -21,3 +21,4 @@ serde = { workspace = true }
 
 [dev-dependencies]
 proptest = { workspace = true }
+serde_json = { workspace = true }
